refactor(phase-3-2): split home_page.py (548→3 files)
拆分 home_page.py (548 行) 为:
- home_page.py (273): HomePage 主类
- _cards.py (279): 4 个卡片类 (GlassCard/QuickCard/ProjectCard/StatChip)
- _empty_state.py (71): EmptyStateCard

验证:
- 6 个 class 完美分布 (1+4+1)
- ruff check all green
- pytest 351 passed, 20 skipped
- 兼容: from .home_page import HomePage 仍工作
